requirements.txt
python-telegram-bot==13.15
yt-dlp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
